+++ test/kb_SetUtilities_server_test.py
@@ -148,6 +148,9 @@ class kb_SetUtilitiesTest(unittest.TestCase):
         self.__class__.genomeName_list[lib_i] = genome_basename
         return new_obj_info
 
+    # DEBUG
+    """
+    
     # call this method to get the WS object info of a Single End Library (will
     # upload the example data if this is the first time the method is called during tests)
     def getSingleEndLibInfo(self, read_lib_basename, lib_i=0):
@@ -357,6 +360,8 @@ class kb_SetUtilitiesTest(unittest.TestCase):
         self.__class__.pairedEndLib_SetInfo = new_obj_set_info
         return new_obj_set_info
 
+    """
+
     ##############
     # UNIT TESTS #
     ##############
@@ -455,7 +460,7 @@ class kb_SetUtilitiesTest(unittest.TestCase):
 
     #### test_KButil_Merge_FeatureSet_Collection_01():
     ##
-    # HIDE @unittest.skip("skipped test_KButil_Merge_FeatureSet_Collection_01()")  # uncomment to skip
+    @unittest.skip("skipped test_KButil_Merge_FeatureSet_Collection_01()")  # uncomment to skip
     def test_KButil_Merge_FeatureSet_Collection_01(self):
         method = 'KButil_Merge_FeatureSet_Collection_01'
         msg = "RUNNING: " + method + "()"
@@ -561,7 +566,7 @@ class kb_SetUtilitiesTest(unittest.TestCase):
 
     #### test_KButil_Slice_FeatureSets_by_Genomes_01():
     ##
-    # HIDE @unittest.skip("skipped test_KButil_Slice_FeatureSets_by_Genomes_01()")  # uncomment to skip
+    @unittest.skip("skipped test_KButil_Slice_FeatureSets_by_Genomes_01()")  # uncomment to skip
     def test_KButil_Slice_FeatureSets_by_Genomes_01(self):
         method = 'KButil_Slice_FeatureSets_by_Genomes_01'
         msg = "RUNNING: " + method + "()"
@@ -645,7 +650,7 @@ class kb_SetUtilitiesTest(unittest.TestCase):
 
     #### test_KButil_Slice_FeatureSets_by_Genomes_NULL_RESULT():
     ##
-    # HIDE @unittest.skip("skipped test_KButil_Slice_FeatureSets_by_Genomes_NULL_RESULT()")  # uncomment to skip
+    @unittest.skip("skipped test_KButil_Slice_FeatureSets_by_Genomes_NULL_RESULT()")  # uncomment to skip
     def test_KButil_Slice_FeatureSets_by_Genomes_NULL_RESULT(self):
         method = 'KButil_Slice_FeatureSets_by_Genomes_NULL_RESULT'
         msg = "RUNNING: " + method + "()"
@@ -726,7 +731,7 @@ class kb_SetUtilitiesTest(unittest.TestCase):
 
     #### test_KButil_Logical_Slice_Two_FeatureSets_01():
     ##
-    # HIDE @unittest.skip("skipped test_KButil_Logical_Slice_Two_FeatureSets_01()")  # uncomment to skip
+    @unittest.skip("skipped test_KButil_Logical_Slice_Two_FeatureSets_01()")  # uncomment to skip
     def test_KButil_Logical_Slice_Two_FeatureSets_01(self):
         method = 'KButil_Logical_Slice_Two_FeatureSets_01'
         msg = "RUNNING: " + method + "()"
@@ -839,7 +844,7 @@ class kb_SetUtilitiesTest(unittest.TestCase):
 
     #### test_KButil_Logical_Slice_Two_FeatureSets_02():
     ##
-    # HIDE @unittest.skip("skipped test_KButil_Logical_Slice_Two_FeatureSets_02()")  # uncomment to skip
+    @unittest.skip("skipped test_KButil_Logical_Slice_Two_FeatureSets_02()")  # uncomment to skip
     def test_KButil_Logical_Slice_Two_FeatureSets_02(self):
         method = 'KButil_Logical_Slice_Two_FeatureSets_02'
         msg = "RUNNING: " + method + "()"
@@ -954,7 +959,7 @@ class kb_SetUtilitiesTest(unittest.TestCase):
 
     #### test_KButil_Logical_Slice_Two_FeatureSets_03():
     ##
-    # HIDE @unittest.skip("skipped test_KButil_Logical_Slice_Two_FeatureSets_03()")  # uncomment to skip
+    @unittest.skip("skipped test_KButil_Logical_Slice_Two_FeatureSets_03()")  # uncomment to skip
     def test_KButil_Logical_Slice_Two_FeatureSets_03(self):
         method = 'KButil_Logical_Slice_Two_FeatureSets_03'
         msg = "RUNNING: " + method + "()"
@@ -1069,7 +1074,7 @@ class kb_SetUtilitiesTest(unittest.TestCase):
 
     #### test_KButil_Logical_Slice_Two_FeatureSets_NULL_RESULT():
     ##
-    # HIDE @unittest.skip("skipped test_KButil_Logical_Slice_Two_FeatureSets_NULL_RESULT()")  # uncomment to skip
+    @unittest.skip("skipped test_KButil_Logical_Slice_Two_FeatureSets_NULL_RESULT()")  # uncomment to skip
     def test_KButil_Logical_Slice_Two_FeatureSets_NULL_RESULT(self):
         method = 'KButil_Logical_Slice_Two_FeatureSets_NULL_RESULT'
         msg = "RUNNING: " + method + "()"
@@ -1180,7 +1185,7 @@ class kb_SetUtilitiesTest(unittest.TestCase):
     
     #### test_KButil_Logical_Slice_Two_AssemblySets_01()
     ##
-    # HIDE @unittest.skip("skipped test_KButil_Logical_Slice_Two_AssemblySets_01()")  # uncomment to skip
+    @unittest.skip("skipped test_KButil_Logical_Slice_Two_AssemblySets_01()")  # uncomment to skip
     def test_KButil_Logical_Slice_Two_AssemblySets_01(self):
         method = 'KButil_Logical_Slice_Two_AssemblySets_01'
         msg = "RUNNING: " + method + "()"
@@ -1283,7 +1288,7 @@ class kb_SetUtilitiesTest(unittest.TestCase):
 
     #### test_KButil_Logical_Slice_Two_GenomeSets_01()
     ##
-    # HIDE @unittest.skip("skipped test_KButil_Logical_Slice_Two_GenomeSets_01()")  # uncomment to skip
+    @unittest.skip("skipped test_KButil_Logical_Slice_Two_GenomeSets_01()")  # uncomment to skip
     def test_KButil_Logical_Slice_Two_GenomeSets_01(self):
         method = 'KButil_Logical_Slice_Two_GenomeSets_01'
         msg = "RUNNING: " + method + "()"
@@ -1381,9 +1386,79 @@ class kb_SetUtilitiesTest(unittest.TestCase):
         pass
 
 
+    #### test_KButil_Split_GenomeSet_01()
+    ##
+    # HIDE @unittest.skip("skipped test_KButil_Split_GenomeSet_01()")  # uncomment to skip
+    def test_KButil_Split_GenomeSet_01(self):
+        method = 'KButil_Split_GenomeSet_01'
+        msg = "RUNNING: " + method + "()"
+        print("\n\n" + msg)
+        print("=" * len(msg) + "\n\n")
+
+        # upload test data
+        # input_data
+        genomeInfo_0 = self.getGenomeInfo('GCF_000287295.1_ASM28729v1_genomic', 0)
+        genomeInfo_1 = self.getGenomeInfo('GCF_000306885.1_ASM30688v1_genomic', 1)
+        genomeInfo_2 = self.getGenomeInfo('GCF_001439985.1_wTPRE_1.0_genomic', 2)
+        genomeInfo_3 = self.getGenomeInfo('GCF_000022285.1_ASM2228v1_genomic', 3)
+
+        genome_ref_0 = self.getWsName() + '/' + str(genomeInfo_0[0]) + '/' + str(genomeInfo_0[4])
+        genome_ref_1 = self.getWsName() + '/' + str(genomeInfo_1[0]) + '/' + str(genomeInfo_1[4])
+        genome_ref_2 = self.getWsName() + '/' + str(genomeInfo_2[0]) + '/' + str(genomeInfo_2[4])
+        genome_ref_3 = self.getWsName() + '/' + str(genomeInfo_3[0]) + '/' + str(genomeInfo_3[4])
+
+        # feature_id_0 = 'A355_RS00030'   # F0F1 ATP Synthase subunit B
+        # feature_id_1 = 'WOO_RS00195'    # F0 ATP Synthase subunit B
+        # feature_id_2 = 'AOR14_RS04755'  # F0 ATP Synthase subunit B
+        # feature_id_3 = 'WRI_RS01560'    # F0 ATP Synthase subunit B
+
+        # GenomeSet 1
+        genomeSet_obj_1 = {'description': 'test genomeSet 1',
+                           'elements': {'genome_0a': {'ref': genome_ref_0},
+                                        'genome_1a': {'ref': genome_ref_1},
+                                        'genome_2a': {'ref': genome_ref_2},
+                                        'genome_3a': {'ref': genome_ref_3}
+                                        }
+                           }
+        provenance = [{}]
+        genomeSet_info = self.getWsClient().save_objects({
+            'workspace': self.getWsName(),
+            'objects': [
+                {
+                    'type': 'KBaseSearch.GenomeSet',
+                    'data': genomeSet_obj_1,
+                    'name': 'test_genomeSet_1',
+                    'meta': {},
+                    'provenance': provenance
+                }
+            ]})[0]
+        set_1_ref = str(genomeSet_info[WSID_I]) + '/' + str(
+            genomeSet_info[OBJID_I]) + '/' + str(genomeSet_info[VERSION_I])
+            
+        # run method
+        num_genome_elements = [2, 1, 1]
+        split_num = 3
+        params = {
+            'workspace_name': self.getWsName(),
+            'input_ref': set_1_ref,
+            'split_num': split_num
+        }
+        result = self.getImpl().KButil_Split_GenomeSet(self.getContext(), params)
+        print('RESULT:')
+        pprint(result)
+
+        # check the output
+        objects_created = result['objects_created']
+        for obj_i,this_obj in enumerate(objects_created):
+            output_ref = this_obj['ref']
+            output_obj = self.getWsClient().get_objects2({'objects': [{'ref': output_ref}]})['data'][0]['data']
+            self.assertEqual(len(list(output_obj['elements'].keys())), num_genome_elements[obj_i])
+        pass
+
+
     #### test_KButil_Merge_GenomeSets_01():
     ##
-    # HIDE @unittest.skip("skipped test_KButil_Merge_GenomeSets_01()")  # uncomment to skip
+    @unittest.skip("skipped test_KButil_Merge_GenomeSets_01()")  # uncomment to skip
     def test_KButil_Merge_GenomeSets_01(self):
         method = 'KButil_Merge_GenomeSets_01'
         msg = "RUNNING: " + method + "()"
@@ -1481,7 +1556,7 @@ class kb_SetUtilitiesTest(unittest.TestCase):
 
     #### test_KButil_Build_GenomeSet_01():
     ##
-    # HIDE @unittest.skip("skipped test_KButil_Build_GenomeSet_01()")  # uncomment to skip
+    @unittest.skip("skipped test_KButil_Build_GenomeSet_01()")  # uncomment to skip
     def test_KButil_Build_GenomeSet_01(self):
         method = 'KButil_Build_GenomeSet_01'
         msg = "RUNNING: " + method + "()"
@@ -1533,7 +1608,7 @@ class kb_SetUtilitiesTest(unittest.TestCase):
 
     #### test_KButil_Build_GenomeSet_from_FeatureSet_01():
     ##
-    # HIDE @unittest.skip("skipped test_KButil_Build_GenomeSet_from_FeatureSet_01()")  # uncomment to skip
+    @unittest.skip("skipped test_KButil_Build_GenomeSet_from_FeatureSet_01()")  # uncomment to skip
     def test_KButil_Build_GenomeSet_from_FeatureSet_01(self):
         method = 'KButil_Build_GenomeSet_from_FeatureSet_01'
         msg = "RUNNING: " + method + "()"
@@ -1616,7 +1691,7 @@ class kb_SetUtilitiesTest(unittest.TestCase):
 
     #### test_KButil_Add_Genomes_to_GenomeSet_01():
     ##
-    # HIDE @unittest.skip("skipped test_KButil_Add_Genomes_to_GenomeSet_01()")  # uncomment to skip
+    @unittest.skip("skipped test_KButil_Add_Genomes_to_GenomeSet_01()")  # uncomment to skip
     def test_KButil_Add_Genomes_to_GenomeSet_01(self):
         method = 'KButil_Add_Genomes_to_GenomeSet_01'
         msg = "RUNNING: " + method + "()"
@@ -1690,7 +1765,7 @@ class kb_SetUtilitiesTest(unittest.TestCase):
 
     #### test_KButil_Add_Genomes_to_GenomeSet_02():
     ##
-    # HIDE @unittest.skip("skipped test_KButil_Add_Genomes_to_GenomeSet_02()")  # uncomment to skip
+    @unittest.skip("skipped test_KButil_Add_Genomes_to_GenomeSet_02()")  # uncomment to skip
     def test_KButil_Add_Genomes_to_GenomeSet_02(self):
         method = 'KButil_Add_Genomes_to_GenomeSet_02'
         msg = "RUNNING: " + method + "()"
@@ -1784,7 +1859,7 @@ class kb_SetUtilitiesTest(unittest.TestCase):
 
     #### test_KButil_Remove_Genomes_from_GenomeSet_01():
     ##
-    # HIDE @unittest.skip("skipped test_KButil_Add_Genomes_to_GenomeSet_01()")  # uncomment to skip
+    @unittest.skip("skipped test_KButil_Add_Genomes_to_GenomeSet_01()")  # uncomment to skip
     def test_KButil_Remove_Genomes_from_GenomeSet_01(self):
         method = 'KButil_Remove_Genomes_from_GenomeSet_01'
         msg = "RUNNING: " + method + "()"
@@ -1862,7 +1937,7 @@ class kb_SetUtilitiesTest(unittest.TestCase):
 
     #### test_KButil_Build_ReadsSet_01()
     ##
-    # HIDE @unittest.skip("skipped test_KButil_Build_ReadsSet_01()")  # uncomment to skip
+    @unittest.skip("skipped test_KButil_Build_ReadsSet_01()")  # uncomment to skip
     def test_KButil_Build_ReadsSet_01(self):
         method = 'KButil_Build_ReadsSet_01'
         msg = "RUNNING: " + method + "()"
@@ -1909,7 +1984,7 @@ class kb_SetUtilitiesTest(unittest.TestCase):
 
     #### test_KButil_Merge_MultipleReadsSets_to_OneReadsSet_01()
     ##
-    # HIDE @unittest.skip("skipped test_KButil_Merge_MultipleReadsSets_to_OneReadsSet_01()")  # uncomment to skip
+    @unittest.skip("skipped test_KButil_Merge_MultipleReadsSets_to_OneReadsSet_01()")  # uncomment to skip
     def test_KButil_Merge_MultipleReadsSets_to_OneReadsSet_01(self):
         method = 'KButil_Merge_MultipleReadsSets_to_OneReadsSet_01'
         msg = "RUNNING: " + method + "()"
@@ -2017,7 +2092,7 @@ class kb_SetUtilitiesTest(unittest.TestCase):
 
     #### test_KButil_Build_AssemblySet_01()
     ##
-    # HIDE @unittest.skip("skipped test_KButil_Build_AssemblySet_01()")  # uncomment to skip
+    @unittest.skip("skipped test_KButil_Build_AssemblySet_01()")  # uncomment to skip
     def test_KButil_Build_AssemblySet_01(self):
         method = 'KButil_Build_AssemblySet_01'
         msg = "RUNNING: " + method + "()"
@@ -2077,7 +2152,7 @@ class kb_SetUtilitiesTest(unittest.TestCase):
 
     #### test_KButil_Batch_Create_ReadsSet_01()
     ##
-    # HIDE @unittest.skip("skipped test_KButil_Batch_Create_ReadsSet_01()")  # uncomment to skip
+    @unittest.skip("skipped test_KButil_Batch_Create_ReadsSet_01()")  # uncomment to skip
     def test_KButil_Batch_Create_ReadsSet_01(self):
         method = 'KButil_Batch_Create_ReadsSet_01'
         msg = "RUNNING: " + method + "()"
@@ -2124,7 +2199,7 @@ class kb_SetUtilitiesTest(unittest.TestCase):
 
     #### test_KButil_Batch_Create_ReadsSet_02()
     ##
-    # HIDE @unittest.skip("skipped test_KButil_Batch_Create_ReadsSet_02()")  # uncomment to skip
+    @unittest.skip("skipped test_KButil_Batch_Create_ReadsSet_02()")  # uncomment to skip
     def test_KButil_Batch_Create_ReadsSet_02(self):
         method = 'KButil_Batch_Create_ReadsSet_02'
         msg = "RUNNING: " + method + "()"
@@ -2178,7 +2253,7 @@ class kb_SetUtilitiesTest(unittest.TestCase):
 
     #### test_KButil_Batch_Create_ReadsSet_03()
     ##
-    # HIDE @unittest.skip("skipped test_KButil_Batch_Create_ReadsSet_03()")  # uncomment to skip
+    @unittest.skip("skipped test_KButil_Batch_Create_ReadsSet_03()")  # uncomment to skip
     def test_KButil_Batch_Create_ReadsSet_03(self):
         method = 'KButil_Batch_Create_ReadsSet_03'
         msg = "RUNNING: " + method + "()"
@@ -2232,7 +2307,7 @@ class kb_SetUtilitiesTest(unittest.TestCase):
 
     #### test_KButil_Batch_Create_AssemblySet_01()
     ##
-    # HIDE @unittest.skip("skipped test_KButil_Batch_Create_AssemblySet_01()")  # uncomment to skip
+    @unittest.skip("skipped test_KButil_Batch_Create_AssemblySet_01()")  # uncomment to skip
     def test_KButil_Batch_Create_AssemblySet_01(self):
         method = 'KButil_Batch_Create_AssemblySet_01'
         msg = "RUNNING: " + method + "()"
@@ -2295,7 +2370,7 @@ class kb_SetUtilitiesTest(unittest.TestCase):
 
     #### test_KButil_Batch_Create_GenomeSet_01()
     ##
-    # HIDE @unittest.skip("skipped test_KButil_Batch_Create_GenomeSet_01()")  # uncomment to skip
+    @unittest.skip("skipped test_KButil_Batch_Create_GenomeSet_01()")  # uncomment to skip
     def test_KButil_Batch_Create_GenomeSet_01(self):
         method = 'KButil_Batch_Create_GenomeSet_01'
         msg = "RUNNING: " + method + "()"
@@ -2353,7 +2428,7 @@ class kb_SetUtilitiesTest(unittest.TestCase):
 
     #### test_KButil_Summarize_GenomeSet_01()
     ##
-    # HIDE @unittest.skip("skipped test_KButil_Summarize_GenomeSet_01()")  # uncomment to skip
+    @unittest.skip("skipped test_KButil_Summarize_GenomeSet_01()")  # uncomment to skip
     def test_KButil_Summarize_GenomeSet_01(self):
         method = 'KButil_Summarize_GenomeSet_01'
         msg = "RUNNING: " + method + "()"