+++ RELEASE_NOTES.md
@@ -1,6 +1,7 @@
 ### Version 1.8.0
 __Changes__
 - Added "Summarize Genome Set" App
+- Added GenomeSet as valid filter to Slice FeatureSets by Genomes()
 
 ### Version 1.7.6
 __Changes__